Demo Site Middleware Redirects: refactor redirects into own file
